% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{cjpwr}
\alias{cjpwr}
\title{Simple Power Analysis and Sample Size Diagnostic for Conjoint Designs}
\usage{
cjpwr(n = 100, t, a = 2, c)
}
\arguments{
\item{n}{The sample size (defaults to 100 for when you're only interested in finding out necessary value of n and don't know what to specify).}

\item{t}{The number of choice-tasks per respondent.}

\item{a}{The number of alternatives per choice task (defaults to two).}

\item{c}{The number of analysis cells - equal to largest number of possible levels for any one feature, or the largest product of levels of any two attributes for power of two-way interaction estimates (Johnson and Orme, 2003).}
}
\description{
Johnson's rule-of-thumb calculation for determining power of conjoint designs.
}
\details{
\code{cjpwr} divides the product of n, t, and a by c, to give Johnson's rule-of-thumb estimation of conjoint design power. It returns a dataframe containing the inputs and result of this calculation, whether (yes/no) this exceeds the minimal minimum threshold (500) and ideal minimum threshold (1000), and the sample sizes (rounded up) necessary for minimum and ideal power thresholds.
}
\examples{
#conjoint design with five choice tasks per respondent, two alternative profiles per task, and maximum six levels per feature
cjpwr(n = 1000, t = 5, a = 2, c = 6)
#same design but considering interactions, where largest product of levels of any two attributes is 18 (6*3)
cjpwr(n = 1000, t = 5, a = 2, c = 18)
#without argument labels, order only matters for c
cjpwr(5, 1000, 2, 18)
cjpwr(5, c = 18, 1000, 2)
cjpwr(c = 18, n = 1000, a = 2, t = 5)
#with defaults
cjpwr(t = 5, c = 18)
}
\seealso{
\code{\link{pwr_n}}
}
