% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpwr_data.R
\name{cjpwr_data}
\alias{cjpwr_data}
\title{Simple Power Analysis and Sample Size Diagnostic for Conjoint Designs}
\usage{
cjpwr_data(data, formula, id, contest_no, profile_no)
}
\arguments{
\item{data}{A tidy, long-formate conjoint dataframe.}

\item{formula}{A formula like those passed to cj/amce/mm functions in `cregg`. RHS variables are used to determine max feature levels.}

\item{id}{A variable, within data, containing respondent IDs. Must be numeric.}

\item{contest_no}{A variable, within data, containing contest/choice-task numbers. Must be numeric.}

\item{profile_no}{A variable, within data, containing within-contest/choice-task profile numbers. Must be numeric.}
}
\description{
Johnson's rule-of-thumb calculation for determining power of conjoint designs.
}
\details{
\code{cjpwr_data} finds and calculates n, t, a, and c from a tidy conjoint data input (with certain variables specified) and divides the product of n, t, and a by c, to give Johnson's rule-of-thumb estimation of conjoint design power. It returns a dataframe containing the inputs and result of this calculation, whether (yes/no) this exceeds the minimal minimum threshold (500) and ideal minimum threshold (1000), and the sample sizes (rounded up) necessary for minimum and ideal power thresholds. {cjpwr_data} uses features of tidyeval which mean variable names can be specified without quoting ("") and without referring back to the dataframe every time (via data$).
}
\examples{
#load example datasets from {cregg} (Leeper 2019)
library(cregg)
data(immigration)
data(taxes)
#run cjpwr_data seamlessly on immigration dataset
#pre-specify formula
f1 <- ChosenImmigrant ~ Gender + Education + LanguageSkills +
CountryOfOrigin + Job + JobExperience + JobPlans +
ReasonForApplication + PriorEntry
cjpwr_data(immigration, f1, CaseID, profile, contest_no)
##for taxes dataset, need to create contest_no and profile_no vars
#work out how many contests/tasks per respondent
tasks_taxes <- sum(taxes$chose_plan)/length(unique(taxes$ID))
#work out number of profiles per task
profiles_taxes <- nrow(taxes$ID)/length(unique(taxes$ID))
profiles_tasks <- profiles_taxes/tasks_taxes
#create profile_no and contest_no based on values of profiles_tasks and tasks_taxes
taxes$profile_no <- rep(c(1,2))
taxes$contest_no <- rep(c(1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8))
#run cjpwr_data for power analysis
cjpwr_data(taxes, chose_plan ~ taxrate1 + taxrate2 + taxrate3 +
taxrate4 + taxrate5 + taxrate6 + taxrev, ID, profile_no, contest_no)
#include interaction variable for a pair of features
library(tidyverse)
immigration$ints <- interaction(immigration$Gender, immigration$PriorEntry)
#specify new formula including interaction feature
f2 <- ChosenImmigrant ~ Gender + Education + LanguageSkills +
CountryOfOrigin + Job + JobExperience + JobPlans +
ReasonForApplication + PriorEntry + ints
#then just run cjpwr_data with the new formula
cjpwr_data(immigration, f2, CaseID, profile, contest_no)
}
