% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr}
\alias{pwr}
\title{Simple Power Analysis for Conjoint Designs}
\usage{
pwr(n, t, a, c)
}
\arguments{
\item{n}{The sample size.}

\item{t}{The number of choice-tasks per respondent.}

\item{a}{The number of alternatives per choice task (2 in most designs).}

\item{c}{The number of analysis cells - equal to largest number of possible levels for any one feature, or the largest product of levels of any two attributes for power of two-way interaction estimates (Johnson and Orme, 2003).}
}
\description{
Johnson's rule-of-thumb calculation for determining power of conjoint designs.
}
\details{
\code{cjpwr} divides the product of n, t, and a by c, to give Johnson's rule-of-thumb estimation of conjoint design power. It returns the result of this calculation and statements of whether this exceeds the minimal minimum threshold (500) and ideal minimum threshold (1000).
}
\examples{
#conjoint design with 1000 respondents, five choice tasks per respondent,
pwr(n = 1000, t = 5, a = 2, c = 6)
#same design but considering interactions, where largest product of levels of any two attributes is 18 (6*3)
pwr(n = 1000, t = 5, a = 2, c = 18)
#without argument labels, order only matters for c
pwr(5, 1000, 2, 18)
pwr(5, c = 18, 1000, 2)
pwr(c = 18, n = 1000, a = 2, t = 5)
}
